+++ src/collective/eeafaceted/collectionwidget/vocabulary.py
@@ -1,19 +1,21 @@
 # encoding: utf-8
 
-from Products.CMFPlone.utils import safe_unicode
-from eea.facetednavigation.interfaces import ICriteria
-from eea.facetednavigation.interfaces import IFacetedNavigable
 from collective.behavior.talcondition.interfaces import ITALConditionable
 from collective.behavior.talcondition.utils import evaluateExpressionFor
 from collective.eeafaceted.collectionwidget.interfaces import ICollectionCategories
 from collective.eeafaceted.collectionwidget.widgets.widget import CollectionWidget
+from eea.facetednavigation.interfaces import ICriteria
+from eea.facetednavigation.interfaces import IFacetedNavigable
+from imio.helpers.cache import get_cachekey_volatile
 from plone import api
+from plone.memoize import ram
+from Products.CMFPlone.utils import safe_unicode
 from zope.component import getAdapter
+from zope.globalrequest import getRequest
 from zope.interface import implements
 from zope.schema.interfaces import IVocabularyFactory
 from zope.schema.vocabulary import SimpleTerm
 from zope.schema.vocabulary import SimpleVocabulary
-from zope.globalrequest import getRequest
 
 
 class CollectionVocabulary(object):
@@ -121,3 +123,32 @@ class CollectionCategoryVocabulary(object):
 
 
 CollectionCategoryVocabularyFactory = CollectionCategoryVocabulary()
+
+
+class CachedCollectionVocabulary(CollectionVocabulary):
+
+    def __call___cachekey(method, self, context):
+        '''cachekey method for self.__call__.'''
+        return self._cache_invalidation_key(context)
+
+    def _cache_invalidation_key(self, context):
+        '''The key will rely on :
+           - current user, in case faceted is stored in the user personal folder;
+           - a stored cache volatile that is destroyed if a DashboardCollection is modified somewhere;
+           - the first facetednavigable context encountered when ascending context parents
+             (useful when collections are defined in a single folder but displayed on various faceted container).'''
+        user = api.user.get_current()
+        date = get_cachekey_volatile('collective.eeafaceted.collectionwidget.cachedcollectionvocabulary')
+        parent = context
+        while not IFacetedNavigable.providedBy(parent) and parent.meta_type != 'Plone Site':
+            parent = parent.aq_parent
+        return user, date, parent
+
+    @ram.cache(__call___cachekey)
+    def __call__(self, context):
+        """Same behaviour as the original CollectionVocabulary, just cached."""
+        terms = super(CachedCollectionVocabulary, self).__call__(context)
+        return terms
+
+
+CachedCollectionVocabularyFactory = CachedCollectionVocabulary()